Update logic of OnSystemContext for widget windows
Updated logic for changing of the
system context for widget windows. For a some
system context types SDL is getting application
pointer not by appId from notification but
just a currently active application.
It works properly when notification is adressed
to the main window, but it works not as expected
for the widget windows. If notification contains
window id then system context change should be
addressed to the application by appId, but not
for a currently active.